LPS-96040 move
