+++ modules/apps/portal-vulcan/portal-vulcan-api/bnd.bnd
@@ -4,7 +4,6 @@ Bundle-Version: 4.0.0
 Export-Package:\
 	com.liferay.portal.vulcan.accept.language,\
 	com.liferay.portal.vulcan.fields,\
-	com.liferay.portal.vulcan.fields.nested,\
 	com.liferay.portal.vulcan.graphql.servlet,\
 	com.liferay.portal.vulcan.multipart,\
 	com.liferay.portal.vulcan.pagination,\
