+++ modules/apps/portal-vulcan/portal-vulcan-api/src/main/java/com/liferay/portal/vulcan/fields/NestedField.java
@@ -12,28 +12,20 @@
  * details.
  */
 
-package com.liferay.portal.vulcan.internal.fields.nested;
+package com.liferay.portal.vulcan.fields;
 
-import com.liferay.petra.lang.CentralizedThreadLocal;
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
 
 /**
  * @author Ivica Cardic
  */
-public class NestedFieldsContextThreadLocal {
+@Retention(RetentionPolicy.RUNTIME)
+@Target(ElementType.METHOD)
+public @interface NestedField {
 
-	public static NestedFieldsContext getNestedFieldsContext() {
-		return _nestedContextThreadLocal.get();
-	}
-
-	public static void setNestedFieldsContext(
-		NestedFieldsContext nestedFieldsContext) {
-
-		_nestedContextThreadLocal.set(nestedFieldsContext);
-	}
-
-	private static final ThreadLocal<NestedFieldsContext>
-		_nestedContextThreadLocal = new CentralizedThreadLocal<>(
-			NestedFieldsContextThreadLocal.class +
-				"._nestedFieldsContextThreadLocal");
+	public String value();
 
 }
