--- modules/apps/portal-vulcan/portal-vulcan-api/src/main/resources/com/liferay/portal/vulcan/fields/nested/packageinfo
@@ -1,1 +1,0 @@
-version 1.0.0
