+++ modules/apps/portal-vulcan/portal-vulcan-impl/src/test/java/com/liferay/portal/vulcan/internal/jaxrs/container/request/filter/NestedFieldsContainerRequestFilterTest.java
@@ -14,8 +14,8 @@
 
 package com.liferay.portal.vulcan.internal.jaxrs.container.request.filter;
 
-import com.liferay.portal.vulcan.internal.fields.nested.NestedFieldsContext;
-import com.liferay.portal.vulcan.internal.fields.nested.NestedFieldsContextThreadLocal;
+import com.liferay.portal.vulcan.internal.fields.NestedFieldsContext;
+import com.liferay.portal.vulcan.internal.fields.NestedFieldsContextThreadLocal;
 
 import java.io.IOException;
